+++ src/stringifiers/dialects/pg.js
@@ -2,7 +2,8 @@ import {
   keysetPagingSelect,
   offsetPagingSelect,
   interpretForOffsetPaging,
-  interpretForKeysetPaging
+  interpretForKeysetPaging,
+  generateCastExpressionFromValueType
 } from '../shared'
 
 const dialect = module.exports = {
@@ -45,8 +46,11 @@ const dialect = module.exports = {
   },
 
   handleBatchedManyToManyPaginated: async function(parent, node, context, tables, batchScope, joinCondition) {
+    const thisKeyOperand = generateCastExpressionFromValueType(
+      `"${node.junction.as}"."${node.junction.sqlBatch.thisKey.name}"`, batchScope[0]
+    )
     const pagingWhereConditions = [
-      `"${node.junction.as}"."${node.junction.sqlBatch.thisKey.name}" = temp."${node.junction.sqlBatch.parentKey.name}"`
+      `${thisKeyOperand} = temp."${node.junction.sqlBatch.parentKey.name}"`
     ]
     if (node.junction.where) {
       pagingWhereConditions.push(
@@ -62,7 +66,7 @@ const dialect = module.exports = {
     const tempTable = `FROM (VALUES ${batchScope.map(val => `(${val})`)}) temp("${node.junction.sqlBatch.parentKey.name}")`
     tables.push(tempTable)
     const lateralJoinCondition =
-      `"${node.junction.as}"."${node.junction.sqlBatch.thisKey.name}" = temp."${node.junction.sqlBatch.parentKey.name}"`
+      `${thisKeyOperand} = temp."${node.junction.sqlBatch.parentKey.name}"`
 
     const lateralJoinOptions = { joinCondition: lateralJoinCondition, joinType: 'LEFT' }
     if (node.where || node.orderBy) {
@@ -157,8 +161,9 @@ const dialect = module.exports = {
   },
 
   handleBatchedOneToManyPaginated: async function(parent, node, context, tables, batchScope) {
+    const thisKeyOperand = generateCastExpressionFromValueType(`"${node.as}"."${node.sqlBatch.thisKey.name}"`, batchScope[0])
     const pagingWhereConditions = [
-      `"${node.as}"."${node.sqlBatch.thisKey.name}" = temp."${node.sqlBatch.parentKey.name}"`
+      `${thisKeyOperand} = temp."${node.sqlBatch.parentKey.name}"`
     ]
     if (node.where) {
       pagingWhereConditions.push(
@@ -167,7 +172,7 @@ const dialect = module.exports = {
     }
     const tempTable = `FROM (VALUES ${batchScope.map(val => `(${val})`)}) temp("${node.sqlBatch.parentKey.name}")`
     tables.push(tempTable)
-    const lateralJoinCondition = `"${node.as}"."${node.sqlBatch.thisKey.name}" = temp."${node.sqlBatch.parentKey.name}"`
+    const lateralJoinCondition = `${thisKeyOperand} = temp."${node.sqlBatch.parentKey.name}"`
     if (node.sortKey) {
       const { limit, order, whereCondition: whereAddendum } = interpretForKeysetPaging(node, dialect)
       pagingWhereConditions.push(whereAddendum)