+++ src/stringifiers/shared.js
@@ -1,9 +1,23 @@
 import { filter } from 'lodash'
 import { cursorToOffset } from 'graphql-relay'
 import { wrap, cursorToObj, maybeQuote } from '../util'
+import idx from 'idx'
 
 export function joinPrefix(prefix) {
   return prefix.slice(1).map(name => name + '__').join('')
+}
+
+
+export function generateCastExpressionFromValueType(key, val) {
+  const castTypes = {
+    string: 'TEXT'
+  }
+  const type = castTypes[typeof val] || null
+
+  if (type) {
+    return `CAST(${key} AS ${type})`
+  }
+  return key
 }
 
 function doubleQuote(str) {
@@ -196,9 +210,9 @@ function sortKeyToWhereCondition(keyObj, descending, sortTable, dialect) {
     sortValues.push(maybeQuote(keyObj[key], name))
   }
   const operator = descending ? '<' : '>'
-  return name === 'oracle' ?
-    recursiveWhereJoin(sortColumns, sortValues, operator) :
-    `(${sortColumns.join(', ')}) ${operator} (${sortValues.join(', ')})`
+  return name === 'oracle'
+    ? recursiveWhereJoin(sortColumns, sortValues, operator)
+    : `(${sortColumns.join(', ')}) ${operator} (${sortValues.join(', ')})`
 }
 
 function recursiveWhereJoin(columns, values, op) {
@@ -215,4 +229,3 @@ function _recursiveWhereJoin(columns, values, op, condition) {
   condition = `(${column} ${op} ${value} OR (${column} = ${value} AND ${condition}))`
   return _recursiveWhereJoin(columns, values, op, condition)
 }
-