+++ test-api/data/schema/setup.js
@@ -6,10 +6,11 @@ const path = require('path')
 const Promise = require('bluebird')
 
 module.exports = function(db, name) {
+  const { ORACLE_URL, PG_URL, MYSQL_URL } = process.env
 
   if (db === 'oracle') {
     console.log('building oracle')
-    const [ password, connectString ] = process.env.ORACLE_URL.split('@')
+    const [ password, connectString ] = ORACLE_URL.split('@')
     const knex = require('knex')({
       client: 'oracledb',
       connection: {
@@ -22,10 +23,7 @@ module.exports = function(db, name) {
 
     let schema = fs.readFileSync(path.join(__dirname, 'oracle.sql')).toString()
     schema = schema.split(/\r?\n\r?\n/)
-    return Promise.mapSeries(schema.filter(i => i), stmt => {
-      console.log(stmt.trim())
-      return knex.raw(stmt.trim())
-    })
+    return Promise.each(schema.filter(i => i), stmt => knex.raw(stmt.trim()))
     .then(() => knex)
     .catch(err => {
       console.error(err)
@@ -35,28 +33,32 @@ module.exports = function(db, name) {
   }
 
   if (db === 'pg') {
-    assert(process.env.PG_URL, 'Must provide environment variable PG_URL, e.g. "postgres://user:pass@localhost/"')
-    const out = execSync(`psql ${process.env.PG_URL + name} < ${__dirname}/pg.sql`)
+    assert(PG_URL, 'Must provide environment variable PG_URL, e.g. "postgres://user:pass@localhost/"')
+    const out = execSync(`psql ${PG_URL + name} < ${__dirname}/pg.sql`)
     if (out.toString()) {
       console.log(out.toString())
     }
     return require('knex')({
       client: 'pg',
-      connection: process.env.PG_URL + name
+      connection: PG_URL + name
     })
   }
 
   if (db === 'mysql') {
-    assert(process.env.MYSQL_URL, 'Must provide environment variable MYSQL_URL, e.g. "mysql://user:pass@localhost/"')
-    const args = parseConnectionStr(process.env.MYSQL_URL + name)
-    const out = execSync(`cat ${__dirname}/mysql.sql | mysql ${args}`)
-    if (out.toString()) {
-      console.log(out.toString())
-    }
-    return require('knex')({
+    assert(MYSQL_URL, 'Must provide environment variable MYSQL_URL, e.g. "mysql://user:pass@localhost/"')
+    const knex = require('knex')({
       client: 'mysql',
-      connection: process.env.MYSQL_URL + name
+      connection: MYSQL_URL + name
     })
+    const ddl = fs.readFileSync(`${__dirname}/mysql.sql`, 'utf8')
+      .split(';')
+      .map(stmt => stmt.trim())
+      .filter(stmt => !!stmt)
+    for (let stmt of ddl) {
+      console.log(stmt)
+    }
+    return Promise.each(ddl, stmt => knex.raw(stmt))
+      .then(() => knex)
   }
 
   if (db === 'sqlite3') {
@@ -75,27 +77,5 @@ module.exports = function(db, name) {
   }
 
   throw new Error(`do not recognize database "${db}"`)
-}
-
-function parseConnectionStr(dbUri) {
-  const parsed = url.parse(dbUri)
-  let args = ''
-  if (parsed.auth) {
-    const [ user, password ] = parsed.auth.split(':')
-    args += '--user=' + user
-    if (password) {
-      args += ' --password=' + password
-    }
-  }
-  if (parsed.hostname) {
-    args += ' --host=' + parsed.hostname
-  }
-  if (parsed.port) {
-    args += ' --port=' + parsed.port
-  }
-  if (parsed.pathname) {
-    args += ` ${parsed.pathname.slice(1)}`
-  }
-  return args
 }