+++ test-api/schema-paginated/QueryRoot.js
@@ -15,6 +15,7 @@ import knex from './database'
 import { User, UserConnection } from './User'
 import Sponsor from './Sponsor'
 import { nodeField } from './Node'
+import ContextPost from './ContextPost'
 
 import joinMonster from '../../src/index'
 import dbCall from '../data/fetch'
@@ -26,7 +27,7 @@ const options = {
   minify: process.env.MINIFY == 1
 }
 if (knex.client.config.client === 'mysql') {
-  options.dialect = PAGINATE ? 'mariadb' : 'mysql'
+  options.dialect = PAGINATE ? 'mysql8' : 'mysql'
 } else if (knex.client.config.client === 'pg') {
   options.dialect = 'pg'
 } else if (knex.client.config.client === 'oracledb') {
@@ -112,7 +113,17 @@ export default new GraphQLObjectType({
           .catch(done)
         }, options)
       }
+    },
+    contextPosts: {
+      type: new GraphQLList(ContextPost),
+        resolve: (parent, args, context, resolveInfo) => {
+          // use the callback version this time
+          return joinMonster(resolveInfo, context, (sql, done) => {
+            knex.raw(sql)
+              .then(data => done(null, data))
+              .catch(done)
+          }, options)
+        }
     }
   })
 })
-